🐛 Erreur de redirection aprés l'effacement d'une évaluation dont on a filtré les restitutions
