+++ app/admin/evaluation_evapro.rb
@@ -83,6 +83,7 @@ ActiveAdmin.register EvaluationEvapro do
   controller do
     include Fichier
     include Admin::DashboardHelper
+    include Admin::EvaluationHelper
 
     helper_method :restitution_globale, :parties,
                   :restitution_pour_situation, :statistiques,
@@ -121,17 +122,6 @@ ActiveAdmin.register EvaluationEvapro do
                 filename: nom_fichier(resource.debutee_le, resource.beneficiaire.nom, "pdf"),
                 type: "application/pdf",
                 disposition: Rails.env.development? ? "inline" : "attachment")
-    end
-
-    def destroy
-      location = if request.referer&.include?("?")
-        request.referer
-      elsif request.referer&.include?(admin_beneficiaire_path(resource.beneficiaire))
-        request.referer
-      else
-        admin_evaluations_evapro_path
-      end
-      destroy!(location: location)
     end
 
     before_action only: :show do
